fix: missing DTOs

+++ src/main/java/com/academia/loja_accenture/modulos/pedido/dto/ProdutoDTO.java
@@ -1,12 +1,13 @@
 package com.academia.loja_accenture.modulos.pedido.dto;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public record ProdutoDTO (
         Long id,
         String nome,
         String descricao,
-        double valor,
+        BigDecimal valor,
         LocalDateTime createdAt,
         Long vendedorId
 ) {
